Type changes, addition of unfinished constructor to Bzet4.

+++ PyBzet4.cpp
@@ -6,11 +6,11 @@ DLLEXPORT Bzet4* Bzet4_newEmpty() {
     return new Bzet4();
 }
 
-DLLEXPORT Bzet4* Bzet4_new(long long bit) {
+DLLEXPORT Bzet4* Bzet4_new(int64_t bit) {
     return new Bzet4(bit);
 }
 
-DLLEXPORT Bzet4* Bzet4_newRange(long long startbit, long long len) {
+DLLEXPORT Bzet4* Bzet4_newRange(int64_t startbit, int64_t len) {
     return new Bzet4(startbit, len);
 }
 
@@ -46,23 +46,23 @@ DLLEXPORT Bzet4* Bzet4_XOR(Bzet4* b, Bzet4* right) {
     return new Bzet4(*b ^ *right);
 }
 
-DLLEXPORT void Bzet4_RANGE(Bzet4* b, long long start, long long len) {
+DLLEXPORT void Bzet4_RANGE(Bzet4* b, int64_t start, int64_t len) {
     b->setRange(start, len);
 }
 
-DLLEXPORT void Bzet4_SET(Bzet4* b, long long bit) {
+DLLEXPORT void Bzet4_SET(Bzet4* b, int64_t bit) {
     b->set(bit);
 }
 
-DLLEXPORT void Bzet4_UNSET(Bzet4* b, long long bit) {
+DLLEXPORT void Bzet4_UNSET(Bzet4* b, int64_t bit) {
     b->unset(bit);
 }
 
-DLLEXPORT bool Bzet4_TEST(Bzet4* b, long long bit) {
+DLLEXPORT bool Bzet4_TEST(Bzet4* b, int64_t bit) {
     return b->at(bit);
 }
 
-DLLEXPORT void Bzet4_FLIP(Bzet4* b, long long bit) {
+DLLEXPORT void Bzet4_FLIP(Bzet4* b, int64_t bit) {
     if (b->at(bit))
         b->unset(bit);
     else
@@ -73,7 +73,7 @@ DLLEXPORT int Bzet4_LEV(Bzet4* b) {
     return b->depth();
 }
 
-DLLEXPORT long long Bzet4_size(Bzet4* b) {
+DLLEXPORT int64_t Bzet4_size(Bzet4* b) {
     return b->size();
 }
 
@@ -89,15 +89,15 @@ DLLEXPORT void Bzet4_HEX(Bzet4* b, int stdOffset) {
     b->printBzet(stdOffset);
 }
 
-DLLEXPORT long long Bzet4_FIRST(Bzet4* b) {
+DLLEXPORT int64_t Bzet4_FIRST(Bzet4* b) {
     return b->firstBit();
 }
 
-DLLEXPORT long long Bzet4_LAST(Bzet4* b) {
+DLLEXPORT int64_t Bzet4_LAST(Bzet4* b) {
     return b->lastBit();
 }
 
-DLLEXPORT long long Bzet4_COUNT(Bzet4* b) {
+DLLEXPORT int64_t Bzet4_COUNT(Bzet4* b) {
     return b->count();
 }
